+++ Source/EbsdLib/HKL/H5CtfVolumeReader.cpp
@@ -116,9 +116,9 @@ int H5CtfVolumeReader::loadData(float* euler1s,
                                 float* euler3s,
                                 int* phases,
                                 bool* goodVoxels,
-                                int xpoints,
-                                int ypoints,
-                                int zpoints,
+                                int64_t xpoints,
+                                int64_t ypoints,
+                                int64_t zpoints,
                                 Ebsd::RefFrameZDir ZDir,
                                 std::vector<QualityMetricFilter::Pointer> filters)
 {
